+++ app/seeds/products.py
@@ -1,5 +1,10 @@
 from app.models import db, Product, environment, SCHEMA
 from sqlalchemy.sql import text
+from sqlalchemy import inspect
+
+def table_exists(table_name):
+    insp = inspect(db.engine)
+    return insp.has_table(table_name, schema=SCHEMA)
 
 def seed_products():
     product1 = Product(
@@ -36,7 +41,7 @@ def seed_products():
     db.session.commit()
 
 def undo_products():
-    if environment == "production":
+    if environment == "production" and table_exists('products'):
         db.session.execute(f"TRUNCATE table {SCHEMA}.products RESTART IDENTITY CASCADE;")
     else:
         db.session.execute(text("DELETE FROM products"))
