+++ app/seeds/users.py
@@ -1,5 +1,10 @@
 from app.models import db, User, environment, SCHEMA
 from sqlalchemy.sql import text
+from sqlalchemy import inspect
+
+def table_exists(table_name):
+    insp = inspect(db.engine)
+    return insp.has_table(table_name, schema=SCHEMA)
 
 # Adds a demo user, you can add other users here if you want
 def seed_users():
@@ -23,7 +28,7 @@ def seed_users():
 # sqlite3 in development you need to instead use DELETE to remove all data and
 # it will reset the primary keys for you as well.
 def undo_users():
-    if environment == "production":
+    if environment == "production" and table_exists('users'):
         db.session.execute(f"TRUNCATE table {SCHEMA}.users RESTART IDENTITY CASCADE;")
     else:
         db.session.execute(text("DELETE FROM users"))
